new update

+++ src/Pixel/Economy/core/Main.php
@@ -61,17 +61,28 @@ class Main extends PluginBase
                             if ($args[0] == "balance") {
                                 if (count($args) == 1) {
                                     $player->sendMessage(TextFormat::GREEN . "Your account balance is: " . TextFormat::BLUE . (string)$this->getter->getBalance($player->getName()) . "$");
-                                } else {
+                                } else if (count($args) == 2) {
                                     $name = $args[1];
 
-                                    $player->sendMessage(TextFormat::GREEN . "Account Balance of. " . $name . " is: " . TextFormat::BLUE . (string)$this->getter->getBalance($name) . "$");
+                                    if ($this->getter->accountExists($name))
+                                        $player->sendMessage(TextFormat::GREEN . "Account Balance of. " . $name . " is: " . TextFormat::BLUE . (string)$this->getter->getBalance($name) . "$");
+                                    else
+                                        $player->sendMessage(TextFormat::RED . $name . " has no bank account!");
+                                } else {
+                                    $player->sendMessage($help);
                                 }
+                            } else if ($args[0] == "help") {
+                                $player->sendMessage($help);
                             } else if ($args[0] == "delete" && count($args) == 2) {
 
                                 $name = $args[1];
 
-                                $this->getter->delete($name);
-                                $player->sendMessage(TextFormat::RED . "You deleted " . $name . " bank account!");
+                                if ($this->getter->accountExists($name)) {
+                                    $this->getter->delete($name);
+                                    $player->sendMessage(TextFormat::RED . "You deleted " . $name . " bank account!");
+                                } else {
+                                    $player->sendMessage(TextFormat::RED . $name . " has no bank account!");
+                                }
                             } else if ($args[0] == "reset") {
                                 $this->getter->reset();
                             } else {
@@ -85,6 +96,15 @@ class Main extends PluginBase
                                 $name = $args[1];
                                 $amount = (int) $args[2];
 
+                                if (!$this->getter->accountExists($name)) {
+                                    $player->sendMessage(TextFormat::RED . $name . " has no bank account!");
+                                    return true;
+                                }
+                                if (!is_numeric($amount)) {
+                                    $player->sendMessage(TextFormat::RED . $amount . " is not a number!");
+                                    return true;
+                                }
+
                                 $this->getter->pay($player->getName(), $name, $amount);
                                 $player->sendMessage(TextFormat::YELLOW . "You payed " . TextFormat::BLUE . $amount . TextFormat::YELLOW . "$ to " . $name);
                             } else if ($args[0] == "set") {
@@ -92,12 +112,30 @@ class Main extends PluginBase
                                 $name = $args[1];
                                 $amount = (int) $args[2];
 
+                                if (!$this->getter->accountExists($name)) {
+                                    $player->sendMessage(TextFormat::RED . $name . " has no bank account!");
+                                    return true;
+                                }
+                                if (!is_numeric($amount)) {
+                                    $player->sendMessage(TextFormat::RED . $amount . " is not a number!");
+                                    return true;
+                                }
+
                                 $this->getter->set($name, $amount);
                                 $player->sendMessage(TextFormat::RED . "You have set the account balance of " . $name . " to " . TextFormat::BLUE . $amount . "$");
                             } else if ($args[0] == "remove") {
 
                                 $name = $args[1];
                                 $amount = (int) $args[2];
+
+                                if (!$this->getter->accountExists($name)) {
+                                    $player->sendMessage(TextFormat::RED . $name . " has no bank account!");
+                                    return true;
+                                }
+                                if (!is_numeric($amount)) {
+                                    $player->sendMessage(TextFormat::RED . $amount . " is not a number!");
+                                    return true;
+                                }
 
                                 $this->getter->remove($name, $amount);
                                 $player->sendMessage(TextFormat::RED . "You have removed " . TextFormat::BLUE . $amount . "$" . TextFormat::RED . " from " . $name);
@@ -117,9 +155,45 @@ class Main extends PluginBase
 
                 if ($cmd->getName() == "money") {
                     if (count($args) == 0) {
+                        $player->sendMessage($help);
                     } else if (count($args) > 0 && count($args) <= 3) {
+                        if ($args[0] == "help") {
+                            $player->sendMessage($help);
+                        } else if ($args[0] == "pay" && count($args) == 3) {
+
+                            $name = $args[1];
+                            $amount = (int) $args[2];
+
+                            if (!$this->getter->accountExists($name)) {
+                                $player->sendMessage(TextFormat::RED . $name . " has no bank account!");
+                                return true;
+                            }
+                            if (!is_numeric($amount)) {
+                                $player->sendMessage(TextFormat::RED . $amount . " is not a number!");
+                                return true;
+                            }
+
+                            $this->getter->pay($player->getName(), $name, $amount);
+                            $player->sendMessage(TextFormat::YELLOW . "You payed " . TextFormat::BLUE . $amount . TextFormat::YELLOW . "$ to " . $name);
+                        } else if ($args[0] == "balance") {
+
+                            if (count($args) == 1) {
+                                $player->sendMessage(TextFormat::GREEN . "Your account balance is: " . TextFormat::BLUE . (string)$this->getter->getBalance($player->getName()) . "$");
+                            } else if (count($args) == 2) {
+                                $name = $args[1];
+
+                                if ($this->getter->accountExists($name))
+                                    $player->sendMessage(TextFormat::GREEN . "Account Balance of. " . $name . " is: " . TextFormat::BLUE . (string)$this->getter->getBalance($name) . "$");
+                                else
+                                    $player->sendMessage(TextFormat::RED . $name . " has no bank account!");
+                            } else {
+                                $player->sendMessage($help);
+                            }
+                        } else {
+                            $player->sendMessage($help);
+                        }
                     } else {
-                        $sender->sendMessage($nocmd);
+                        $player->sendMessage($help);
                     }
                 }
 
@@ -133,11 +207,9 @@ class Main extends PluginBase
 
         /**************  CONSOLE  **************/
         else {
+            if ($cmd->getName() == "money") {
+            }
         }
-
-        if ($cmd->getName() == "money") {
-        }
-
 
         return true;
     }
